fix(dev): Prime mtimes cache
When watching for file/directory changes, prime the FileChangeTracker
with initial mtimes of the tracked files, such that if the first
notification received for a file is spurious (does not represent an
actual change), shouldEmitChange will return 'false'.

Without this, the 'priming' happened via several reloads/restarts, but
with the introduction of the ForkPool, the FileChangeTracker cache no
longer perists across restarts, never allowing it to become populated.
Result: unpredictable restarts.